feat(meta-service): Add comprehensive snapshot database metrics (#18422)
Add detailed snapshot database metrics to monitor storage performance and statistics:
- Block count, data size, index size metrics
- Average block size and keys per block calculations
- Block read statistics (total, from cache, from disk)
- Upgrade rotbl dependency to 0.2.7 for enhanced access statistics

These metrics provide visibility into snapshot storage efficiency and cache performance.

All metrics are `gauge` type:

```config
# number of keys in the last snapshot.
metasrv_server_snapshot_key_count 0

# number of primary keys in the last snapshot.
metasrv_server_snapshot_primary_index_count 0

# number of expire index keys in the last snapshot.
metasrv_server_snapshot_expire_index_count 0

# number of blocks in the last snapshot.
metasrv_server_snapshot_block_count 0

# size of data section in the last snapshot.
metasrv_server_snapshot_data_size 0

# size of index section in the last snapshot.
metasrv_server_snapshot_index_size 0

# average size of a block in the last snapshot.
metasrv_server_snapshot_avg_block_size 0

# average number of keys per block in the last snapshot.
metasrv_server_snapshot_avg_keys_per_block 0

# total number of read block from cache or from disk.
metasrv_server_snapshot_read_block 0

# total number of read block from cache.
metasrv_server_snapshot_read_block_from_cache 0

# total number of read block from disk.
metasrv_server_snapshot_read_block_from_disk 0
```

+++ src/meta/service/src/api/grpc/grpc_service.rs
@@ -577,7 +577,8 @@ impl MetaService for MetaServiceImpl {
                 wal_closed_chunk_sizes: status.raft_log.wal_closed_chunk_sizes,
             }),
 
-            snapshot_key_count: status.snapshot_key_count as u64,
+            snapshot_key_count: status.snapshot_key_count,
+
             state: status.state,
             is_leader: status.is_leader,
             current_term: status.current_term,
